Alter some dictionary operations to work in Py3
dict.keys() returns a view object in py3. we can iterate over them, but
since we want to have the keys sorted as well, we have to turn the view
into a list and sort them via the sorted function.

Use in operator instead of dict.has_key() to test for key membership as
it works in py2.2+.