Add Quick Start to README
Adapted from an answer to "Graphing a process's memory usage" on [Stack Overflow](https://stackoverflow.com/a/62876993/111424).

Memory profiler is exactly the tool I've been looking for. But the Usage section of the README makes it seem so complicated to use that it scared me away!

The Stack Overflow answer gets straight to the point and shows how simple it is to use. I think it deserves to be at the top of the README.